backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
google-generativeai==0.3.2
python-multipart==0.0.9
typing-extensions==4.9.0
langchain==0.1.9
langchain-core==0.1.27
tenacity==8.2.3
ipywidgets>=8.0.0
jupyter>=1.0.0
crewai
litellm 
google-cloud-aiplatform
crewai-tools
streamlit
flask
